requirements.txt
Flask
flask-cors
flasgger
APScheduler
pytz
gunicorn
requests
psycopg2-binary